Upgraded to Java 1.7 Upgraded RXTX library to newer implementation Set project version to 0.2

+++ freebus-fts-knxcomm/src/main/java/org/freebus/knxcomm/application/GroupValueResponse.java
@@ -32,7 +32,7 @@ public final class GroupValueResponse extends GenericDataApplication
     * 
     * @param data - the data.
     */
-   public GroupValueResponse(int[] data)
+   public GroupValueResponse(byte[] data)
    {
       super(ApplicationType.GroupValue_Response, data);
    }
